Move lambdas into methods
