+++ src/miner.h
@@ -18,6 +18,7 @@
 
 #include <boost/multi_index_container.hpp>
 #include <boost/multi_index/ordered_index.hpp>
+#include <masternodes/mn_checks.h>
 
 class CBlockIndex;
 class CChainParams;
@@ -140,6 +141,41 @@ struct update_for_parent_inclusion
     CTxMemPool::txiter iter;
 };
 
+typedef std::array<std::uint8_t, 20> EvmAddress;
+
+struct EvmAddressWithNonce {
+    EvmAddress address;
+    uint64_t nonce;
+
+    bool operator<(const EvmAddressWithNonce& item) const {
+        return std::tie(address, nonce) < std::tie(item.address, item.nonce);
+    }
+};
+
+struct EvmPackageContext {
+    // Used to track EVM TXs by sender and nonce.
+    std::map<EvmAddressWithNonce, uint64_t> feeMap;
+    // Used to track EVM nonce and TXs by sender
+    std::map<EvmAddress, std::map<uint64_t, CTxMemPool::txiter>> addressTxsMap;
+    // Keep track of EVM entries that failed nonce check
+    std::multimap<uint64_t, CTxMemPool::txiter> failedNonces;
+    // Used for replacement Eth TXs when a TX in chain pays a higher fee
+    std::map<uint64_t, CTxMemPool::txiter> replaceByFee;
+};
+ 
+struct EvmTxPreApplyContext {
+    const CTransaction& tx;
+    CTxMemPool::txiter& txIter;
+    std::vector<unsigned char>& txMetadata;
+    CustomTxType txType;
+    int height;
+    uint64_t evmQueueId;
+    EvmPackageContext& pkgCtx;
+    std::set<uint256>& checkedTXEntries;
+    CTxMemPool::setEntries& inBlockEntries;
+    CTxMemPool::setEntries& failedTxEntries;
+};
+
 /** Generate a new block, without valid proof-of-work */
 class BlockAssembler
 {
@@ -219,6 +255,7 @@ private:
     void RemoveTxs(const std::set<uint256> &txHashSet, const std::map<uint256, CAmount> &txFees);
     /** Remove specific TX from the block */
     void RemoveTxIters(const std::vector<CTxMemPool::txiter> iters);
+    bool EvmTxPreapply(const EvmTxPreApplyContext& ctx);
 };
 
 /** Modify the extranonce in a block */
